os/kernel/mqueue: Assertion fix incase of non zero in nwaitnotempty and no task in wait list
The assertion is that under certain circumstances, there's an inconsistency between:
1. The count of waiting tasks (msgq->nwaitnotempty)
2. The actual tasks found in the wait list (g_watingformqnotempty)
This inconsistency can happen when task is terminated (via pthread_cancel or similar)
while waiting on a message queue. If the task is removed from the waitlist
but the counter isn't properly decremented, we end up with assert sutuation.

This fix address the  assert by checking the senario to resets the counter to zero.

Signed-off-by: prabu Rajendran <prabu.r@samsung.com>