os/kernel/task: cleaning up message queue wait state when a task terminates.
Under certian circumstances, there's an inconsistency between:
1. The count of waiting tasks(msgq->nwaitnotempty)
2. The actual tasks found in the wait list (g_waitingformqnotempty)

This inconsistency can happen when a task is terminated (via pthread_cancel or similar)
while waiting on a message queue. If the task is removed from the waitlist
but the counter isn't properly decremented, we endup with a assert situation.

Fix : This addresses the root cause by properly cleaning up the appropriate counters of the message
queue when a task terminates.

Signed-off-by: prabu Rajendran <prabu.r@samsung.com>